+++ CHANGELOG.md
@@ -13,6 +13,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Added sort-by support via GUI.
 - Added similarity selection option.
 - Show confusion matrix in evaluation results for object detection.
+- Added option to select annotation source for class balancing.
 
 ### Changed